+++ react/hooks/useSimilarProducts/index.ts
@@ -1,0 +1,2 @@
+export * from './useSimilarProducts';
+export * from './useSimilarProducts.types';