+++ react/hooks/useSimilarProducts/useSimilarProducts.types.ts
@@ -1,0 +1,12 @@
+import type { ProductTypes } from 'vtex.product-context';
+
+interface SimilarProductsContextType {
+  similarProducts: ProductTypes.Product[];
+}
+
+interface SimilarProductsProviderProps {
+  children: React.ReactNode;
+  showUnavailable?: boolean;
+}
+
+export { SimilarProductsContextType, SimilarProductsProviderProps };